feat: ✨ auto height with page header
