+++ apps/builder/src/middleware/reduxAsync/sendMessage/componentsMethod.ts
@@ -168,7 +168,6 @@ export const componentsAsync = (
       break
     }
     case "setComponentPropsReducer":
-    case "updateComponentPropsReducerNotWithUndoRedo":
     case "updateComponentPropsReducer":
       const updatePayload: UpdateComponentPropsPayload = payload
       const finalNode = searchDsl(
