+++ apps/builder/src/page/App/components/DotPanel/components/Canvas/interface.ts
@@ -6,6 +6,7 @@ export interface RenderComponentCanvasContainerProps {
   canResizeCanvas?: boolean
   safeRowNumber?: number
   handleUpdateHeight?: (height: number) => void
+  minHeight?: number
 }
 
 export interface DropCollectedProps {
