+++ apps/builder/src/page/App/components/DotPanel/components/Section/BodySection/interface.ts
@@ -1,0 +1,3 @@
+import { RenderBasicSectionProps } from "../interface"
+
+export type RenderSectionProps = RenderBasicSectionProps
