+++ apps/builder/src/page/App/components/DotPanel/components/Section/BodySection/style.ts
@@ -1,0 +1,9 @@
+import { css } from "@emotion/react"
+
+export const bodySectionWrapperStyle = css`
+  position: absolute;
+  width: var(--illa-canvas-body-width, 100%);
+  left: var(--illa-canvas-body-left, 0);
+  top: var(--illa-canvas-body-top, 0);
+  height: var(--illa-canvas-body-height);
+`
