+++ apps/builder/src/page/App/components/DotPanel/components/Section/ChangeLayoutBar/HorizontalLayoutBar/index.tsx
@@ -1,4 +1,4 @@
-import { FC, useCallback } from "react"
+import { FC, memo, useCallback } from "react"
 import { useDispatch } from "react-redux"
 import { ReactComponent as ChangeLayout } from "@/assets/change-layout.svg"
 import { componentsActions } from "@/redux/currentApp/editor/components/componentsSlice"
@@ -6,20 +6,14 @@ import {
   PageNodeProps,
   SECTION_POSITION,
 } from "@/redux/currentApp/editor/components/componentsState"
-import {
-  ChangeVerticalLayoutBarProps,
-  changeHorizontalLayoutBarProps,
-} from "./interface"
+import { changeHorizontalLayoutBarProps } from "./interface"
 import {
   changeHorizontalLayoutBarWrapperStyle,
   changeHorizontalLayoutLeftIconStyle,
   changeLayoutHorizontalBarStyle,
-  changeLayoutVerticalBarStyle,
-  changeVerticalLayoutBarWrapperStyle,
-  changeVerticalLayoutLeftIconStyle,
 } from "./style"
 
-export const ChangeHorizontalLayoutBar: FC<changeHorizontalLayoutBarProps> = (
+const ChangeHorizontalLayoutBar: FC<changeHorizontalLayoutBarProps> = (
   props,
 ) => {
   const {
@@ -89,55 +83,6 @@ export const ChangeHorizontalLayoutBar: FC<changeHorizontalLayoutBarProps> = (
   )
 }
 
-export const ChangeVerticalLayoutBar: FC<ChangeVerticalLayoutBarProps> = (
-  props,
-) => {
-  const { direction, currentPosition, currentPageName, targetSectionName } =
-    props
+ChangeHorizontalLayoutBar.displayName = "ChangeHorizontalLayoutBar"
 
-  const dispatch = useDispatch()
-
-  const changeAction = useCallback(() => {
-    const newProps: Partial<PageNodeProps> = {
-      layout: "Custom",
-    }
-
-    switch (currentPosition) {
-      case SECTION_POSITION.TOP:
-      case SECTION_POSITION.BOTTOM: {
-        if (targetSectionName === "leftSection") {
-          newProps.leftPosition = SECTION_POSITION.FULL
-        } else {
-          newProps.rightPosition = SECTION_POSITION.FULL
-        }
-        break
-      }
-      case SECTION_POSITION.CENTER: {
-        if (targetSectionName === "leftSection") {
-          newProps.leftPosition =
-            direction === "top" ? SECTION_POSITION.TOP : SECTION_POSITION.BOTTOM
-        } else {
-          newProps.rightPosition =
-            direction === "top" ? SECTION_POSITION.TOP : SECTION_POSITION.BOTTOM
-        }
-        break
-      }
-    }
-    dispatch(
-      componentsActions.updateTargetPagePropsReducer({
-        pageName: currentPageName,
-        newProps: newProps,
-      }),
-    )
-  }, [currentPageName, currentPosition, direction, dispatch, targetSectionName])
-
-  return (
-    <div
-      css={changeVerticalLayoutBarWrapperStyle(direction)}
-      onClick={changeAction}
-    >
-      <ChangeLayout css={changeVerticalLayoutLeftIconStyle(direction)} />
-      <div css={changeLayoutVerticalBarStyle} />
-    </div>
-  )
-}
+export default memo(ChangeHorizontalLayoutBar)
