+++ apps/builder/src/page/App/components/DotPanel/components/Section/ChangeLayoutBar/HorizontalLayoutBar/interface.ts
@@ -1,0 +1,6 @@
+import { ChangeLayoutBarProps } from "../interface"
+
+export interface changeHorizontalLayoutBarProps extends ChangeLayoutBarProps {
+  sectionName: "headerSection" | "footerSection"
+  direction: "left" | "right"
+}
