+++ apps/builder/src/page/App/components/DotPanel/components/Section/ChangeLayoutBar/VerticalLayoutBar/interface.ts
@@ -1,0 +1,6 @@
+import { ChangeLayoutBarProps } from "../interface"
+
+export interface ChangeVerticalLayoutBarProps extends ChangeLayoutBarProps {
+  direction: "top" | "bottom"
+  sectionName: "leftSection" | "rightSection"
+}
