+++ apps/builder/src/page/App/components/DotPanel/components/Section/ChangeLayoutBar/interface.ts
@@ -1,0 +1,13 @@
+import { SECTION_POSITION } from "@/redux/currentApp/editor/components/componentsState"
+
+export interface ChangeLayoutBarProps {
+  direction: "top" | "bottom" | "left" | "right"
+  currentPosition: SECTION_POSITION
+  currentPageName: string
+  sectionName:
+    | "leftSection"
+    | "rightSection"
+    | "headerSection"
+    | "footerSection"
+  targetSectionName: "leftSection" | "rightSection"
+}
