+++ apps/builder/src/page/App/components/DotPanel/components/Section/ModalSection/interface.ts
@@ -1,0 +1,6 @@
+import { ModalSectionNode } from "@/redux/currentApp/editor/components/componentsState"
+
+export interface RenderModalSectionProps {
+  sectionNode: ModalSectionNode
+  columnNumber: number
+}
