+++ apps/builder/src/page/App/components/DotPanel/components/Section/ModalSection/style.ts
@@ -1,0 +1,10 @@
+import { css } from "@emotion/react"
+
+export const modalWrapperStyle = css`
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  z-index: 8;
+`
