+++ apps/builder/src/page/App/components/DotPanel/components/Section/index.ts
@@ -1,0 +1,6 @@
+export { RenderBodySection } from "./BodySection"
+export { RenderModalSection } from "./ModalSection"
+export { RenderHeaderSection } from "./HeaderSection"
+export { RenderFooterSection } from "./FooterSection"
+export { RenderLeftSection } from "./LeftSection"
+export { RenderRightSection } from "./RightSection"
