+++ apps/builder/src/page/App/components/ScaleSquare/components/DragContainer/index.tsx
@@ -83,6 +83,7 @@ export const DragContainer: FC<DragContainerProps> = (props) => {
       displayName,
       unitWidth,
       columnNumber,
+      canDrag,
     ],
   )
