+++ apps/builder/src/redux/currentApp/editor/components/componentsReducer.ts
@@ -214,21 +214,6 @@ export const setComponentPropsReducer: CaseReducer<
   node.props = getNewWidgetPropsByUpdateSlice(updateSlice, {})
 }
 
-export const updateComponentPropsReducerNotWithUndoRedo: CaseReducer<
-  ComponentsState,
-  PayloadAction<UpdateComponentPropsPayload>
-> = (state, action) => {
-  const { displayName, updateSlice } = action.payload
-  if (!isObject(updateSlice) || !displayName) {
-    return
-  }
-  const node = searchDsl(state, displayName)
-  if (!node) return
-  const widgetProps = node.props || {}
-  const clonedWidgetProps = cloneDeep(widgetProps)
-  node.props = getNewWidgetPropsByUpdateSlice(updateSlice, clonedWidgetProps)
-}
-
 export const updateMultiComponentPropsReducer: CaseReducer<
   ComponentsState,
   PayloadAction<UpdateComponentPropsPayload[]>
