+++ apps/builder/src/redux/currentApp/editor/components/componentsSlice.ts
@@ -23,7 +23,6 @@ import {
   updateComponentLayoutInfoReducer,
   updateComponentNodeHeightReducer,
   updateComponentPropsReducer,
-  updateComponentPropsReducerNotWithUndoRedo,
   updateComponentReflowReducer,
   updateMultiComponentPropsReducer,
   updateRootNodePropsReducer,
@@ -43,7 +42,6 @@ const componentsSlice = createSlice({
     initComponentReducer,
     updateComponentPropsReducer,
     setComponentPropsReducer,
-    updateComponentPropsReducerNotWithUndoRedo,
     deleteComponentNodeReducer,
     sortComponentNodeChildrenReducer,
     updateComponentDisplayNameReducer,
