+++ apps/builder/src/redux/currentApp/executionTree/executionListener.ts
@@ -465,7 +465,6 @@ export function setupExecutionListeners(
         componentsActions.addModalComponentReducer,
         componentsActions.setGlobalStateReducer,
         componentsActions.deleteGlobalStateByKeyReducer,
-        componentsActions.updateComponentPropsReducerNotWithUndoRedo,
         actionActions.addActionItemReducer,
         actionActions.removeActionItemReducer,
         actionActions.updateActionItemReducer,
