+++ apps/builder/src/widgetLibrary/ContainerWidget/emptyState.tsx
@@ -1,20 +1,20 @@
-import { FC, ReactNode } from "react"
+import { FC, useEffect } from "react"
 import { useTranslation } from "react-i18next"
 import {
   applyEmptyStateWrapperStyle,
   emptyStateStyle,
-  keyPressComponentStyle,
 } from "@/widgetLibrary/ContainerWidget/style"
+import { UNIT_HEIGHT } from "../../page/App/components/DotPanel/constant/canvas"
 
-interface IKeyPressComponentProps {
-  children?: ReactNode
-}
-export const KeyPressComponent: FC<IKeyPressComponentProps> = (props) => {
-  return <span css={keyPressComponentStyle}>{props.children}</span>
-}
-
-export const ContainerEmptyState: FC = () => {
+export const ContainerEmptyState: FC<{
+  handleUpdateHeight?: (height: number) => void
+}> = ({ handleUpdateHeight }) => {
   const { t } = useTranslation()
+  useEffect(() => {
+    if (handleUpdateHeight) {
+      handleUpdateHeight(15 * UNIT_HEIGHT)
+    }
+  }, [handleUpdateHeight])
   return (
     <div css={applyEmptyStateWrapperStyle}>
       <span css={emptyStateStyle}>{t("widget.container.empty")}</span>
