+++ apps/builder/src/widgetLibrary/PublicSector/TransformWidgetWrapper/renderWithJSON.tsx
@@ -73,25 +73,13 @@ export const TransformWidgetWrapperWithJson: FC<TransformWidgetWrapperWithJsonPr
     )
 
     const handleUpdateOriginalDSLMultiAttr = useCallback(
-      (updateSlice: Record<string, any>) => {
+      (updateSlice: Record<string, any>, notUseUndoRedo?: boolean) => {
         if (!isObject(updateSlice)) return
         dispatch(
           componentsActions.updateComponentPropsReducer({
             displayName: displayName,
             updateSlice,
-          }),
-        )
-      },
-      [dispatch, displayName],
-    )
-
-    const handleUpdateOriginalDSLMultiAttrNotUseUnDoRedo = useCallback(
-      (updateSlice: Record<string, any>) => {
-        if (!isObject(updateSlice)) return
-        dispatch(
-          componentsActions.updateComponentPropsReducerNotWithUndoRedo({
-            displayName: displayName,
-            updateSlice,
+            notUseUndoRedo,
           }),
         )
       },
@@ -99,25 +87,17 @@ export const TransformWidgetWrapperWithJson: FC<TransformWidgetWrapperWithJsonPr
     )
 
     const handleUpdateOriginalDSLOtherMultiAttr = useCallback(
-      (displayName: string, updateSlice: Record<string, any>) => {
+      (
+        displayName: string,
+        updateSlice: Record<string, any>,
+        notUseUndoRedo?: boolean,
+      ) => {
         if (!displayName || !isObject(updateSlice)) return
         dispatch(
           componentsActions.updateComponentPropsReducer({
             displayName,
             updateSlice,
-          }),
-        )
-      },
-      [dispatch],
-    )
-
-    const handleUpdateOriginalDSLOtherMultiAttrNotUseUnDoRedo = useCallback(
-      (displayName: string, updateSlice: Record<string, any>) => {
-        if (!displayName || !isObject(updateSlice)) return
-        dispatch(
-          componentsActions.updateComponentPropsReducerNotWithUndoRedo({
-            displayName,
-            updateSlice,
+            notUseUndoRedo,
           }),
         )
       },
@@ -268,12 +248,6 @@ export const TransformWidgetWrapperWithJson: FC<TransformWidgetWrapperWithJsonPr
           componentNode={componentNode}
           triggerEventHandler={triggerEventHandler}
           triggerMappedEventHandler={triggerMappedEventHandler}
-          handleUpdateOriginalDSLMultiAttrNotUseUnDoRedo={
-            handleUpdateOriginalDSLMultiAttrNotUseUnDoRedo
-          }
-          handleUpdateOriginalDSLOtherMultiAttrNotUseUnDoRedo={
-            handleUpdateOriginalDSLOtherMultiAttrNotUseUnDoRedo
-          }
         />
       </div>
     )
